requirements.txt
africastalking==1.2.4
certifi==2021.5.30
charset-normalizer==2.0.3
contextlib2==21.6.0
dj-database-url==0.5.0
Django==2.1.7
django-stubs==1.8.0
django-stubs-ext==0.2.0
djangorestframework==3.9.2
djangorestframework-stubs==0.3.0
gunicorn==20.1.0
idna==3.2
mypy==0.910
mypy-extensions==0.4.3
psycopg2-binary==2.9.1
pytz==2018.9
requests==2.26.0
schema==0.7.4
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
whitenoise==5.3.0
